Use request pool to limit concurrent image requests
This switches from using an array of Promises to fetch books'
images (where we were trying to resolve all of them at once)
to using a Pool which by default (and we can change this if we
want to) makes 25 concurrent requests.

The unused Api::getPool method was also deleted, to avoid
confusion.

Bug: T225686