Switch to using system fonts
Rather than including our own fonts, instead include fonts from
the host system.

Get the font info with the fc-list command, and cache it for a
week.

Turn the FontProvider into a Symfony service class and inject it
everwhere required. This makes it easier to test, and will make
more sense as we progressively convert more code to the Symfony
architecture.

This patch also makes it possible to provide *any* available
font name via the URL, e.g. `?fonts=WikisourceMono`, even if
the name isn't listed in the dropdown in the export form.

Bug: T261479