Switch to Symfony framework
This patch migrates the entire application to Symfony, and tries to modify the
smallest amount of code possible to make this work.

Some notes:

* Tests in CI were already not passing, so the failing tests are now excluded
(via the `exclude-from-ci` test group). These are all to do with epub standards
conformance, and will be tackled later.
* The CLI entry point bin/book.php has been deleted, which might impact some
people's scripts.
* A bunch of bits of old code have been moved to Controllers, Commands, or
templates, without change. These are sometimes showing as changes in the Git
diff, but actually still need to be brought up to date.
* Fully setting up Doctrine isn't done yet. Although the single DB table we
need could be done via a migration now, it's not yet able to be used with
Doctrine ORM because it doesn't contain a simple primary key, so it seems
easier to leave this for a subsequent patch.
* There is duplication in tests for configuring the database, but this seems
acceptable (rather than refactoring it to somewhere common) because it'll all
be removed after the above.
* Docker Compose operation is complicated now by the need to split environment
variables between `.env` and `.env.local`.
* HTTP status text is not working correctly (but the response code is).

Bug: T263577