Sanitize output filenames of books
Exported filenames should not contain subpage slashes. They do
actually work, at least for some browsers, but Symfony doesn't
allow them and they also mean that books exported on the CLI
are put into subdirectories (which can be a bit odd — this also
makes this a breaking change, but we've already broken the
CLI format with the change to app:export so I don't think that
matters).

This change also removes the return values from BookCreator::create()
and replaces them with other getter methods on that class. This'll
make it easier when we come to turn that class into an injectable
service.

Bug: T263577