Refactor Api::$lang
Move the method dealing with getting the lang from the request
into a controller, and add new accessor method to Api class. This
is a step towards making the Api class into a service.

The main part of this (other than moving things around) is changing
how the default language is determined.

Bug: T265660