Update font-handling for better selection, weights, etc.
The existing font family name handling wasn't taking into account
the fact that fc-list reports these with alternate names appearing
after a comma. This removes this trailing info, and also adds more
thorough weight and slant handling.

The dropdown list is also updated to include all available fonts,
because it sounds like at some point we'll update this to be more
responsive and useful but for now this makes it much easier to
test and see what is available.

Bug: T261479