Remove instantiations of Api class
Remove the last instantiations of the Api class, in preparation
for being able to inject it everywhere (and inject it along with
other services like a cache and HTTP client; that'll come in a
subsequent patch).

The main changes here are to make GeneratorSelector injectable,
and to remove one of the factory methods from BookCreator.

The fact that we have to pass $api to Util::getTempFile() is annoying,
and means that two of the Generator classes have to know about the Api,
but this is a workaround that we can remove once we've changed the
temp file storage to also be injectable.

Bug: T265660