Fix references sometimes opening as external links
I'm unsure about the repercussions of changing how we load the html
content from XML to HTML but the later seems to fix this issue. It
is also how we load pages' content from parsoid.

This issue was happening becuase `$this->xPath->query( '//html:a[contains(@style, "mw-Ref")]' )`
was not working when the DOMDocument content was loaded via loadHTML.
Removing the namespace from the above query fixes the book references but breaks the About
page unless it is loaded as HTML.