Add on-wiki config to define default fonts
Set up an on-wiki config page MediaWiki:WS_Export.json, which
can be used to store any per-Wikisource configuration. So far
this is just `defaultFont` but it can be extended as needed.

Defaults for these configs live in the tool's code, not in
Multilingual Wikisource (like the About page) because that wiki
will need its own config.

The existing default of FreeSerif will need to be migrated to
the on-wiki config once this change has been merged.

Bug: https://phabricator.wikimedia.org/T274561