Improve subpage-finding to not use URLs
This changes the chapter subpage finding logic (when there's no
ws-summary in use) to use the links' title attribute rather than
their href attribute. It also moves the prefix-searching out of
XPath and into the link loop.

The fragment (#) links are no longer excluded because it wasn't
clear why this is ever necessary; it looks like a pre-Parsoid thing.

Bug: https://phabricator.wikimedia.org/T275870