Deny access from IPs globally blocked on Wikimedia wikis
This is to prevent automated disruption, since stewards are pretty on
top of blocking webhosts. Some users with a VPN may get blocked, but
will be able to login to get around it (as well as the preexisting
IP-based rate limiting).

A 'Login' link is not shown in the normal UI, only on the error page,
because this is the only situation where you'd need to login.

Bug: T286048