Throw exception for rejected page requests
This returns to the previous behaviour of throwing a
WsExportException when a wiki page is not found.

Also changes the error message to call it a 'page' rather than
a 'book', becuase it could apply to a subpage or the root-level
page, and it isn't referring to the whole book.

Bug: T280329