Check for imageinfo rather than missing or invalid
Because shared-repo images are reported as missing, do not check
for that status. And to make things simpler, use the same handling
for invalid ones as well, and just check to see if any imageinfo
metadata has been returned (that's all we're interested in, after
all).

This is a follow-up to GH #512.

Bug: T372956